wip(#1101): extract lower/stmt
